+++ nerfbaselines/methods/gaussian_splatting.py
@@ -28,7 +28,7 @@ import tempfile
 import numpy as np
 from PIL import Image
 from nerfbaselines import (
-    Method, MethodInfo, ModelInfo, OptimizeEmbeddingsOutput, RenderOutput, Cameras, camera_model_to_int, Dataset, NoGPUError
+    Method, MethodInfo, ModelInfo, OptimizeEmbeddingsOutput, RenderOutput, Cameras, camera_model_to_int, Dataset
 )
 from nerfbaselines.io import wget
 import shlex
@@ -53,34 +53,6 @@ from utils.loss_utils import l1_loss, ssim  # type: ignore
 from utils.sh_utils import SH2RGB  # type: ignore
 from scene import Scene, sceneLoadTypeCallbacks  # type: ignore
 from utils import camera_utils  # type: ignore
-
-
-def remap_error(fn):
-    def is_gpu_error(e: Exception) -> bool:
-        if isinstance(e, NoGPUError):
-            return True
-        if isinstance(e, RuntimeError):
-            return "Found no NVIDIA driver on your system." in str(e)
-        if isinstance(e, EnvironmentError):
-            return "unknown compute capability. ensure pytorch with cuda support is installed." in str(e).lower()
-        if isinstance(e, ImportError):
-            return "libcuda.so.1: cannot open shared object file" in str(e)
-        return False
-
-    if getattr(fn, "__error_remap__", False):
-        return fn
-
-    @functools.wraps(fn)
-    def wrapped(*args, **kwargs):
-        try:
-            return fn(*args, **kwargs)
-        except Exception as e:
-            if is_gpu_error(e):
-                raise NoGPUError from e
-            raise e
-
-    wrapped.__error_remap__ = True  # type: ignore
-    return wrapped
 
 
 def flatten_hparams(hparams, *, separator: str = "/", _prefix: str = ""):
@@ -271,7 +243,6 @@ def _convert_dataset_to_gaussian_splatting(dataset: Optional[Dataset], tempdir: 
 
 
 class GaussianSplatting(Method):
-    @remap_error
     def __init__(self, *,
                  checkpoint: Optional[str] = None,
                  train_dataset: Optional[Dataset] = None,
