+++ nerfbaselines/methods/nerfw_reimpl.py
@@ -22,7 +22,6 @@ from nerfbaselines import (
     Method, Dataset, Cameras, RenderOutput,
     OptimizeEmbeddingsOutput, 
     ModelInfo, MethodInfo, CameraModel,
-    NoGPUError,
 )
 from nerfbaselines.utils import invert_transform, pad_poses, convert_image_dtype
 from nerfbaselines import cameras as _cameras
@@ -50,34 +49,6 @@ import train  # type: ignore
 
 logger = logging.getLogger("nerfw-reimpl")
 warnings.filterwarnings("ignore", ".*does not have many workers.*")
-
-
-def remap_error(fn):
-    def is_gpu_error(e: Exception) -> bool:
-        if isinstance(e, NoGPUError):
-            return True
-        if isinstance(e, RuntimeError):
-            return "Found no NVIDIA driver on your system." in str(e)
-        if isinstance(e, EnvironmentError):
-            return "unknown compute capability. ensure pytorch with cuda support is installed." in str(e).lower()
-        if isinstance(e, ImportError):
-            return "libcuda.so.1: cannot open shared object file" in str(e)
-        return False
-
-    if getattr(fn, "__error_remap__", False):
-        return fn
-
-    @functools.wraps(fn)
-    def wrapped(*args, **kwargs):
-        try:
-            return fn(*args, **kwargs)
-        except Exception as e:
-            if is_gpu_error(e):
-                raise NoGPUError from e
-            raise e
-
-    wrapped.__error_remap__ = True  # type: ignore
-    return wrapped
 
 
 def flatten_hparams(hparams, *, separator: str = "/", _prefix: str = ""):
@@ -400,7 +371,6 @@ def get_opts(config_overrides):
 
 
 class NeRFWReimpl(Method):
-    @remap_error
     def __init__(self, *,
                  checkpoint: Optional[str] = None,
                  train_dataset: Optional[Dataset] = None,
